+++ src/adapters/react/index.ts
@@ -1,3 +1,9 @@
+export {
+	useCreateQueryClient,
+	useInitQueryClient,
+	useQueryClient,
+	useSafeQueryFn,
+} from "../../core/services/query.service.js";
 export * from "./fetch.js";
 export * from "./query.js";
 export * from "./watch.js";